src/requirements.txt
bidict==0.21.4
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.0.4
Cython==3.0.12
dataclasses==0.8
distro==1.9.0
dnspython==2.2.1
eventlet==0.33.3
Flask==2.0.3
Flask-SocketIO==5.3.6
greenlet==2.0.2
h11==0.13.0
idna==3.10
importlib-metadata==4.8.3
influxdb==5.3.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.5
numpy==1.19.4
packaging==21.3
paho-mqtt==1.6.1
Pillow==8.4.0
pkg_resources==0.0.0
pyparsing==3.1.4
pytesseract==0.3.7
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2025.2
requests==2.27.1
scikit-build==0.16.7
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.1.1
urllib3==1.26.20
uuid==1.30
Werkzeug==2.0.3
wsproto==1.0.0
zipp==3.6.0
